requirements.txt
pandas
langchain
langchain_openai
openpyxl
python-dotenv
tiktoken
pdfplumber